dash_tanstack_pivot/pivot_engine/requirements.txt
# Core Dependencies
duckdb>=0.8.0
pyarrow>=10.0.0
pyarrow_hotfix>=0.4
ibis-framework>=4.0.0; extra != "no-ibis"
pandas>=1.5.0

# API & Web
fastapi>=0.100.0; extra != "no-api"
uvicorn[standard]>=0.20.0; extra != "no-api"
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Security & Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Observability
prometheus-fastapi-instrumentator>=6.0.0
structlog>=23.1.0

# Development & Testing
pytest>=7.0.0
pytest-asyncio>=0.20.0
httpx>=0.24.0
